requirements.txt
alembic==1.4.3
amqp==2.6.1
Babel==2.8.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap4==4.0.2
Flask-DebugToolbar==0.11.0
Flask-Ext==0.1
Flask-Limiter==1.4
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
flaskbb-plugin-proxyfix==1.0
idna==2.10
infinity==1.5
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
limits==1.5.1
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1secure_filename
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
stripe==2.50.0
urllib3==1.25.10
validators==0.18.0
vine==1.3.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4
